pyproject.toml
[project]
name = "ing-zadanie"
version = "0.1.0"
description = ""
authors = [
    {name = "Zenek",email = "gracjanzeniuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-playwright (>=0.7.0,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
